Move AntPatternFileFilter to utils package
- Create dedicated utils package for common/utility classes
- Move AntPatternFileFilter from main package to utils package
- Update import statement in JsonYamlValidatorMojo

Addresses code review feedback for better package organization
