+++ src/main/java/com/dataliquid/maven/plugin/schema/validator/JsonYamlValidatorMojo.java
@@ -24,6 +24,8 @@ import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
+import com.dataliquid.maven.plugin.schema.validator.utils.AntPatternFileFilter;
+
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
